upgrading distance

+++ speed/bike_logic.h
@@ -15,12 +15,11 @@ const float wheel_circumference = 3.1416 * diameter_m; // meters
 const int red_above_this = 35;
 const int segments_per_revolution = 60;
 const float cadence_to_kph_scalar=wheel_circumference*gear_ratio*60/1000; //The 60 is to go from per minute to per hour the 1000 is to go from m to km
-const float changes_per_ten_meters=10/(wheel_circumference/segments_per_revolution); //Used to work out km. Back of envelope is about 300 
+const int changes_per_ten_meters= (int) 10/(wheel_circumference/segments_per_revolution); //Used to work out km. Back of envelope is about 300 
 
 extern int histogram[10];
 extern long total;
 extern int count;
-
 
 
 void printColorStats(const int histogram[], long total, int count);
